+++ ClubInfluApp/BusinessLogic/Interfaces/ICuponServicioService.cs
@@ -6,7 +6,7 @@ namespace ClubInfluApp.BusinessLogic.Interfaces
     public interface ICuponServicioService
     {
         public void ReservarCuponOfertaServicio(int idOfertaServicio);
-
         public List<CuponServicioViewModel> ObtenerCuponesPorOfertaServicio(int idOfertaServicio);
+        public string validarCuponDeServicioPorCodigo(string codigoDeCuponAValidar);
     }
 }
