+++ ClubInfluApp/Data/Interfaces/ICuponServicioRepository.cs
@@ -6,11 +6,9 @@ namespace ClubInfluApp.Data.Interfaces
     public interface ICuponServicioRepository
     {
         public void ReservarCuponOfertaServicio(int idOfertaServicio, int idUsuarioInfluencer);
-
         public string ValidarCuponOfertaServicio(int idOfertaServicio, int idInfluencer);
-
         public OfertaServicioViewModel ObtenetCodigoNombreOfertaPorOfertaServicio(int idOfertaServicio);
-
         public List<CuponServicioViewModel> ObtenerCuponesPorOfertaServicio(int idOfertaServicio);
+        public string validarCuponDeServicioPorCodigo(string codigoDeCuponAValidar);
     }
 }
