+++ ClubInfluApp/Data/Interfaces/IOfertaServicioRepository.cs
@@ -8,6 +8,5 @@ namespace ClubInfluApp.Data.Interfaces
         public List<OfertaServicioViewModel> ObtenerOfertasDeServicioFiltradas(FiltroOfertasDeServicio filtroOfertasDeServicio);
         public List<OfertaServicioViewModel> ObtenerOfertasDeServicioPorEmpresa(int idEmpresa);
         public int CrearOfertaDeServicio(OfertaServicio ofertaServicio);
-        public string validarCuponDeServicioPorCodigo(string codigoDeCuponAValidar);
     }
 }
