+++ ClubInfluApp/Data/Repositories/CuponServicioRepository.cs
@@ -142,6 +142,33 @@ namespace ClubInfluApp.Data.Repositories
                 throw;
             }
         }
+
+        public string validarCuponDeServicioPorCodigo(string codigoDeCuponAValidar)
+        {
+            using NpgsqlConnection connection = new NpgsqlConnection(dbConnectionString);
+            connection.Open();
+
+            try
+            {
+                string validarCuponQuery = "SELECT * FROM validar_cupon_servicio_por_codigo(@p_codigoCuponServicio);";
+
+                string mensajeDeRespuesta = connection
+                    .Query<string>(validarCuponQuery, new
+                    {
+                        p_codigoCuponServicio = codigoDeCuponAValidar
+                    }).FirstOrDefault() ?? "Error en base de datos";
+
+                return mensajeDeRespuesta;
+            }
+            catch
+            {
+                throw;
+            }
+            finally
+            {
+                connection.Close();
+            }
+        }
     }
 }
